Made logic that checks if the rendered day cell is the same day of any of the returned schedules day. If so, render a little highlighted text saying how many workouts at what time on the day's cell.
